refactor(OrderBook): move Summary to a new module
