+++ src/OrderBook/OrderBookSection.js
@@ -1,5 +1,5 @@
 import { Bag } from "./Bag.js";
-import { goToFormPage } from "../main.js";
+import { Summary } from "./Summary.js";
 
 export function OrderBookSection() {
   const section = document.createElement("section");
@@ -12,37 +12,8 @@ export function OrderBookSection() {
   const bag = Bag();
   section.append(bag);
 
-  const summary = document.createElement("div");
-  summary.className = "summary";
-
-  const totalPrice = TotalPrice();
-  summary.append(totalPrice);
-
-  const confirmOrderBtn = ConfirmOrderBtn();
-  summary.append(confirmOrderBtn);
-
+  const summary = Summary();
   section.append(summary);
 
   return section;
-}
-
-function TotalPrice() {
-  const totalPriceElement = document.createElement("p");
-  totalPriceElement.className = "total";
-  totalPriceElement.innerText = " $";
-
-  const totalPriceSpan = document.createElement("span");
-  totalPriceSpan.id = "total-price";
-  totalPriceSpan.innerText = "0";
-  totalPriceElement.prepend(totalPriceSpan);
-
-  return totalPriceElement;
-}
-
-function ConfirmOrderBtn() {
-  const btn = document.createElement("button");
-  btn.innerHTML = "Confirm order";
-  btn.onclick = goToFormPage;
-
-  return btn;
 }
